Добавил возможность использовать внешние методы для маппинга полей

+++ dto/person.go
@@ -6,7 +6,7 @@ type Person struct {
 	FirstName string
 	LastName string
 	MiddleName string
-	Age int64
+	Age int
 	Account external.Account
 	Profile Profile
 	Projects []string
